requirements.txt
asgiref==3.2.7
autopep8==1.5
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
idna==2.9
joblib==0.14.1
numpy==1.18.2
pandas==1.0.3
pycodestyle==2.5.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
